Rename snake_case packages to lowerCamelCase
Resolves #17
