+++ src/main/kotlin/org/treeWare/common/codec/SkipUnknownStateMachine.kt
@@ -1,7 +1,7 @@
 package org.treeWare.common.codec
 
-import org.treeWare.model.codec.decoding_state_machine.AuxDecodingStateMachine
-import org.treeWare.model.codec.decoding_state_machine.DecodingStack
+import org.treeWare.model.codec.decoder.stateMachine.AuxDecodingStateMachine
+import org.treeWare.model.codec.decoder.stateMachine.DecodingStack
 import java.math.BigDecimal
 
 class SkipUnknownStateMachine<Aux>(
