+++ src/main/kotlin/org/treeWare/model/codec/JsonModelEncoder.kt
@@ -1,10 +1,11 @@
 package org.treeWare.model.codec
 
 import org.treeWare.common.codec.JsonWireFormatEncoder
-import org.treeWare.model.codec.aux_encoder.AuxEncoder
+import org.treeWare.common.traversal.TraversalAction
+import org.treeWare.model.codec.encoder.AuxEncoder
+import org.treeWare.model.codec.encoder.ModelEncodingVisitor
 import org.treeWare.model.core.ElementModel
 import org.treeWare.model.operator.forEach
-import org.treeWare.common.traversal.TraversalAction
 import java.io.Writer
 
 fun <Aux> encodeJson(
