+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/ModelDecodingStateMachine.kt
@@ -1,4 +1,4 @@
-package org.treeWare.model.codec.decoding_state_machine
+package org.treeWare.model.codec.decoder.stateMachine
 
 import org.apache.logging.log4j.LogManager
 import org.treeWare.common.codec.DecodingStateMachine
